Day 4

+++ mywebsite/home/views.py
@@ -1,13 +1,44 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.models import User
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 
 # rooms =[
 #     {'id': 1, 'name':'Lets learn python!'},
 #     {'id': 2, 'name':'Design with me'},
 #     {'id': 3, 'name':'Fontend developers'},
 # ] 
+
+
+def loginPage(request):
+    if request.method =='POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User dose not exist")
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request,user)
+            return redirect('index')
+        else:
+            messages.error(request, "Username or password does not exist")
+
+    context = {}
+    return render(request, 'home/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('index')
 
 
 def index(request):
@@ -29,6 +60,8 @@ def room(request, pk):
     context = {'room':room}      
     return render(request,'home/room.html',context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
 
     form = RoomForm()
@@ -42,9 +75,14 @@ def createRoom(request):
     return render(request, 'home/room_form.html', context)
 
 
+
+@login_required(login_url='login')
 def updateRoom(request ,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -55,8 +93,16 @@ def updateRoom(request ,pk):
     context = {'form':form}
     return render(request,'home/room_form.html', context)
 
+
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
+
+
     if request.method =='POST':
         room.delete()
         return redirect('index')
